Slight refactor
Facegen transfer works insconsistenly in MO2 when the Facegen files are in a BSA. For some NPCs the .nif is transferred and for others it's not, even though everything is coming from Skyim - Meshes.bsa. Needs to be figured out and fixed.

Settings autogeneration for PluginsExcludedFromMerge does not currently work.